moving and renaming as requested by professor
