+++ src/game/scenes/PieceStyle.js
@@ -1,0 +1,11 @@
+export class pieceStyle {
+	styleNumber;
+
+	constructor(number) {
+		this.styleNumber = number;
+	}
+
+	changeStyle(number) {
+		this.styleNumber = number;
+	}
+}
